introduced .set(...) method for multivectors and scalars
This allows to avoid certain intermediate variables and thus makes the
generated code a bit shorter at the cost of slightly more complex
generator code.

But these cases should be easily optimizable by subsequent compilation
steps.  So this commit goes to a branch that will probably not be merged
back to the main branch.