another optimization for sandwiches
Multiple purely operator-based terms with the same bitmap
are now added up before being multiplie with some operand component.

The code has become quite complex and there are still possibilities
for further optimization (sharing of sums across operands).
I think it is possible to both simplify and optimize.

It looks like this is heading towards generating a (sparse) matrix
for the operator and applying it to the operand...