optional inlining of sub-expressions
This is not for optimization but for readability of the generated code.

Variable management and the API of the WASM back end had to be modified
(cleaned up?) quite a bit.  For the WebGL back end it was trivial since
both variable names and more complex expressions are just strings.
And the numeric back end ignores the option.

TODO: Make use of this option in more places in the Algebra code.